// src/api/models/biblioteca.models.ts
// Generated by https://quicktype.io

export interface BibliotecaModel {
  id: number;
  nombre: string;
  tipo: string;
  direccion: string;
  codigoPostal: string;
  longitud: string;
  latitud: string;
  telefono: string;
  email: string;
  descripcion: string;
  createdAt?: string;
  updatedAt?: string;
  LocalidadCodigoLocalidad: string;
  Localidad?: LocalidadModel;
}

export interface LocalidadModel {
  codigoLocalidad: string;
  nombreLocalidad: string;
  createdAt?: string;
  updatedAt?: string;
  ProvinciumCodigoProvincia: string;
  Provincium?: ProvinciumModel;
}

export interface ProvinciumModel {
  codigoProvincia: string;
  nombreProvincia: string;
  createdAt?: string;
  updatedAt?: string;
}